refactor(job): changed property type of some value objects

+++ code/automata/src/main/java/com/automata/job/domain/valueobject/ProgramCurrentWideRateDto.java
@@ -1,5 +1,5 @@
 package com.automata.job.domain.valueobject;
 
-public record ProgramCurrentWideRateDto(Long programId, int currentWideRate) {
+public record ProgramCurrentWideRateDto(Long programId, Long currentWideRate) {
 
 }
